+++ swagger_parser/lib/src/generator/config/generator_config.dart
@@ -78,16 +78,17 @@ class GeneratorConfig {
   /// Default content type for all requests and responses.
   ///
   /// If the content type does not match the default, generates:
-  /// @Headers(<String, String>{'Content-Type': 'PARSED CONTENT TYPE'})
+  /// `@Headers(<String, String>{'Content-Type': 'PARSED CONTENT TYPE'})`
   final String defaultContentType;
 
   /// DART ONLY
   /// Add extra parameter to all requests. Supported after retrofit 4.1.0.
   ///
   /// If  value is 'true', then the annotation will be added to all requests.
-  ///
+  /// ```dart
   /// @POST('/path/')
   /// Future<String> myMethod({@Extras() Map<String, dynamic>? extras});
+  /// ```
   final bool extrasParameterByDefault;
 
   /// Optional. Set regex replacement rules for the names of the generated classes/enums.